Setup Commands Rig
The approach we are taking here is to allow each command sit on its own.
If Crystal was a scripting language, each of these commands would be its
own script. Since this is compiled, the best we can do is to separate
the concerns for each command into its own class.

To mediate between this commands, the router uses user input to
determine the commands to run the command. Here, I do some variant of
Ruby's "method_missing" — I'm no master of Crystal yet, so the best I
can do is to approximate. The router suppliments for OptionParser where
we cannot find the command, and runs the command if it's supported by
Crit.

I ran into some issues trying to test this router though; especially the
delegation bit. There's more text on this in the router_spec file. Turns
out there is a lot of room a dynamically typed language allows you play
in.

With this command rig setup, any command can be added to the tool by
defining it as a Crit::SUPPORTED_COMMAND, modifying the Crit::Router's
 #route method, and implementing the appropriate command.